Dodanie usuniętego fragmentu
